HARP-7907: Limit number of raycasts used for Camera tilt calculus.
In some cases focus point (and tilt angle) may not be calculated because
Camera is looking to the void. In this scenarios the last known focus point
is used and the highest possible tilt value is exposed (90 degrees).
This allows to keep camera pivot consistent (when changing the projection
type, or manipulating camera manually) and computations based on tilt
stable (using maximum tilt), this may be: fog, fading effect, etc.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>